tpm commands/responses have 32-bit sizes

+++ tpm2/errors.go
@@ -8,5 +8,5 @@ const (
 
 func (r TPMRC) Error() string {
 	// TODO: Formatting and other TPM error comprehension
-	return fmt.Sprintf("TPM error code: %x", r)
+	return fmt.Sprintf("TPM error code: %x", uint32(r))
 }
